// land-alarmy-nest/src/interface/RTMSDataSvcSHRent.interface.ts
export interface RTMSDataSvcSHRent_Inf {
  useOfRenewalRequestRights: string | undefined;
  yearOfConstruction: string | undefined;
  contractClassification: string | undefined;
  term: string | undefined;
  size: string | undefined;
  year: number | undefined;
  beopjeongDong: string | undefined;
  depositAmount: string | undefined;
  month: number | undefined;
  monthlyRentAmount: number | undefined;
  day: number | undefined;
  previousContractDeposit: number;
  previousContractMonthlyRent: number;
  code: string;
}
// "갱신요구권사용": [" " ],
// "건축년도": ["1997"],
// "계약구분": ["갱신"],
// "계약기간": ["23.10~25.09"],
// "계약면적": ["45"],
// "년": ["2023"],
// "법정동": ["청운동"],
// "보증금액": ["11,550"],
// "월": ["9"],
// "월세금액": [ "21"],
// "일": ["16" ],
// "종전계약보증금": [ "11,000"],
// "종전계약월세": [ "20"],
// "지역코드": [ "11110 ]